Update_PlayerAttackController
자동 공격(클릭) 기능 구현 완료

+++ Assets/Scripts/Player/PlayerSO.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 public class PlayerSO : ScriptableObject
 {
     [SerializeField] public int maxHealth;
-    [SerializeField] public int Damage;
+    [SerializeField] public float Damage;
     [SerializeField] public float evasionRate; // 회피율
     [SerializeField] public float AutoAttackSpeed; //자동 클릭 속도
 }
